// packages/ui-emotion/src/components/Typography/types.tsx
export interface HeadingProps {
  type?: 'heading'
  variant?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
  className?: string
}

export interface BodyProps {
  type?: 'body'
  variant?: 'body1' | 'body2' | 'body3'
  className?: string
}

export type CaptionProps = {
  type?: 'caption'
  variant?: 'caption1' | 'caption2'
  className?: string
}

export type SubTitleProps = {
  type?: 'subTitle'
  variant?: 'subTitle1' | 'subTitle2' | 'subTitle3'
  className?: string
}